refactor(models): export PlanScope, PlanSpecInput, ReviewScope to public API
Add three symbols to models/__init__.py lazy-export registry and replace
all deep imports (from vibe3.models.plan/review import) across src and
tests with public API imports (from vibe3.models import).

Changes:
- models/__init__.py: add PlanScope, PlanSpecInput, ReviewScope to
  TYPE_CHECKING, _LAZY_IMPORTS, and __all__
- roles/plan.py: use public API imports instead of deep imports
- roles/review.py: use public API imports instead of deep imports
- 7 test files: update imports to use public API

This addresses the "blocked imports" identified in commit 096c2999
(refactor(roles): switch deep imports to public APIs).

Issue: #2190
