+++ tests/vibe3/agents/test_run_prompt.py
@@ -9,7 +9,7 @@ from vibe3.agents.run_prompt import (
     build_run_task_section,
 )
 from vibe3.config.settings import VibeConfig
-from vibe3.models.plan import PlanRequest, PlanScope
+from vibe3.models import PlanRequest, PlanScope
 
 FORBIDDEN_INTERNAL_TOKENS = (
     "common.md",
